tests/wpt/tests/css/css-gaps/animation/gap-decorations-color-neutral-keyframe-002.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>gap decorations column-rule-color neutral keyframe</title>
    <link rel="help" href="https://drafts.csswg.org/css-gaps-1/#column-row-rule-color">
    <meta name="assert" content="gap decorations column-rule-color value list supports neutral keyframe.">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/css/support/interpolation-testcommon.js"></script>
  </head>
  <body>
    <div id="target"></div>
    <script>
      test(() => {
        target.style.columnRuleStyle = 'solid';
        target.style.columnRuleColor = 'black black black';
        var animation = target.animate([{}, {columnRuleColor: 'red blue rgb(255, 0, 0)'}], 1000);
        animation.pause();
        animation.currentTime = 500;
        assert_equals(getComputedStyle(target).columnRuleColor, 'rgb(128, 0, 0) rgb(0, 0, 128) rgb(128, 0, 0)');
      }, 'gap decorations column-rule-color value list supports neutral keyframe.');
    </script>
  </body>
</html>
